refactor(board): rename board-set/ to board-sets/
The folder holds the board-sets UI (list + info/delete dialogs); the plural
matches the domain concept and its data-side counterpart (board-sets-store),
making the deliberate UI/data layer split read as intentional rather than an
accidental singular/plural mismatch.
