+++ src/features/board/index.ts
@@ -1,6 +1,6 @@
-export { BoardSetDeleteDialog } from "./board-set/delete-dialog";
-export { BoardSetInfoDialog } from "./board-set/info-dialog";
-export { BoardSetList } from "./board-set/list";
+export { BoardSetDeleteDialog } from "./board-sets/delete-dialog";
+export { BoardSetInfoDialog } from "./board-sets/info-dialog";
+export { BoardSetList } from "./board-sets/list";
 export { BoardViewer } from "./board-viewer";
 export { importBoardFromUrl } from "./import/import-from-url";
 export { useImportBoardFiles } from "./import/use-import-board-files";
